+++ template/public/constants/paths.js
